requirements.txt
django
djangorestframework
django-cors-headers
python-docx
Pillow